requirements.txt
Imath==0.0.2
importlib_metadata==4.11.1
matplotlib==3.5.0
mrcfile==1.3.0
open3d==0.15.1
opencv_contrib_python==4.5.5.62
OpenEXR==1.3.9
Pillow==9.2.0
plyfile==0.7.4
pytorch_fid==0.2.1
scikit_learn==1.1.2
scikit_video==1.1.11
scipy==1.7.3
tabulate==0.8.9
torch_ema==0.3
tqdm==4.62.3
